Silence -Wformat warnings regarding 'int' vs. 'unsigned int' arguments.
* gettext-tools/src/read-catalog-abstract.c (catalog_reader_parse): Use %u
instead of %d in format string.
* gettext-tools/src/read-properties.c (UTF16_VALUE): Cast result to
'unsigned short'.
* gettext-tools/src/write-java.c (write_lookup_code,
write_java2_init_statements, write_java_code): Use %u instead of %d in format
strings.
* gettext-tools/src/x-perl.c (extract_balanced): Cast enum-typed fprintf
argument to 'int'.

+++ gettext-tools/src/x-perl.c
@@ -3607,7 +3607,7 @@ extract_balanced (message_list_ty *mlp,
 
         default:
           fprintf (stderr, "%s:%d: unknown token type %d\n",
-                   real_file_name, tp->line_number, tp->type);
+                   real_file_name, tp->line_number, (int) tp->type);
           abort ();
         }
